BugFixingExplosion
- switched Instantiate()-Method

+++ schmup_prototyp/Assets/DestroyAfterSeconds.cs
@@ -5,13 +5,13 @@ using UnityEngine;
 public class DestroyAfterSeconds : MonoBehaviour {
 
 
-	public int Seconds;
+	public float Seconds;
 
 	// Use this for initialization
 	void Start () {
 		
 		if(Seconds == null)
-			Seconds = 1;
+			Seconds = 1.0f;
 	
 		Destroy(gameObject, Seconds);
 	}
